Initial Github actions test framework
This commit adds a github action where the outputs of several test
'cbrain' commands are compared to an expected set of outputs.

As the command's code changes, edit the two files that store the
commands and their expected outputs.

capture_tests/cbrain_cli_commands : the commands
capture_tests/expected_captures.txt : the outputs

At the end of the workflow, GitHub will prepare an artifact file
with a suitable 'git diff' command output that can be applied to
the 'expected_captures.txt' file using 'git apply'.